% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_rownames}
\alias{add_rownames}
\title{Turn rownames into labelled column for kable}
\usage{
add_rownames(x, label = names(dimnames(x))[1], ...)
}
\description{
Turn rownames into labelled column for kable
}
