% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{decomp_table}
\alias{decomp_table}
\title{Table of gaps, percent gaps from model fitting log salaries}
\usage{
decomp_table(x, log = TRUE, p = TRUE, n_min = 1)
}
\arguments{
\item{z}{output of decomp2 or decomp1 (untested)}
}
\description{
Table of gaps, percent gaps from model fitting log salaries
}
