% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomps.R
\name{decomps}
\alias{decomps}
\title{Gender gap decompositions without SEs}
\usage{
decomps(
  fitl,
  g = "Gender",
  comp = "M",
  data = na.omit(getD(full)),
  cond = NULL,
  refit = TRUE
)
}
\arguments{
\item{fitl}{a named list of models}

\item{g}{character string with the name of the variable used for grouping, e.g. "Gender"}

\item{comp}{character string giving the name of the level used as a comparator, e.g. "M"}

\item{cond}{optional vector with names of variables used for conditioning}

\item{refit}{logical (default TRUE) should models be refit using 'data'}

\item{comon}{data set from which all models can be fit}
}
\value{
an object of class 'decomps' consisting of a list with two data frames: the first, called 'gaps', consists of the
rbind of 'data' for each model
together with additional variables:
\itemize{
  \item res_: ordinary residual
  \item pred_: ordinary predicted value
  \item resp_: response
  \item predg_: value predicted if comparator
  \item resg_: residual from value predicted if comparator
  \item model: identifying the model using 'names(fitl)'
}

The second component, called 'summ', is the mean summary of
of 'gaps' keeping variables indicating each model, the
grouping ('g') and conditioning ('cond') variables, if any.
}
\description{
Gender gap decompositions without SEs
}
