% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmt}
\alias{fmt}
\title{Formatting that thinks it's smart}
\usage{
fmt(
  x,
  nsmall = ns_(x, nsig, nright),
  nsig = 4,
  nright = 1,
  scientific = FALSE,
  big.mark = ",",
  ...
)
}
\arguments{
\item{x}{numeric object to be formatted}

\item{nsmall, }{nsig, nright,}

\item{scientific}{FALSE by default}

\item{big.mark}{is ',' by default}

\item{...}{other arguments to \code{\link{format}}}
}
\description{
Chooses number of significant digits semi-intelligently
}
