% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{icp}
\alias{icp}
\title{Graphical comparison of AIC and BIC}
\usage{
icp(ll, ...)
}
\arguments{
\item{ll}{a list of models using the same data.}

\item{rot}{angle to rotate labels}
}
\description{
Compares AIC and BIC for a list of models. The models should have been fitted
to the same data.
}
