% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kb}
\alias{kb}
\title{Collect options for kbl}
\usage{
kb(x, caption = NULL, format.args = list(big.mark = ","), ...)
}
\arguments{
\item{x}{object to print}

\item{caption}{default: NULL}

\item{format.args}{default: list(big.mark = ',')}

\item{...}{other arguments}
}
\description{
Collect options for kbl
}
