% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{kbind}
\alias{kbind}
\title{Multiple rows for kable}
\usage{
kbind(x, ..., prefix = "", align = "r")
}
\arguments{
\item{x}{element in first row}

\item{...}{additional elements}

\item{prefix}{add before first element}

\item{align}{'r', 'l' or 'c', default 'r'}
}
\description{
Multiple rows for kable
}
