% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{kprint}
\alias{kprint}
\alias{kprint.Gap}
\alias{kprint.lm}
\title{Printing generic function to produce kable output}
\usage{
kprint(object, ...)

\method{kprint}{Gap}(object, formula, type = 1, ...)

\method{kprint}{lm}(object, ...)
}
\arguments{
\item{object}{a Gap object created by \code{\link{Gap}}}

\item{...}{other arguments}
}
\description{
Printing generic function to produce kable output

Printing Gap object for kable output

Printing lm object for kable output
}
\section{Methods (by class)}{
\itemize{
\item \code{kprint(Gap)}: method for Gap objects

\item \code{kprint(lm)}: method for printing lm output

}}
