% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{lhs}
\alias{lhs}
\title{Extract the LHS of a formula as a character}
\usage{
lhs(fmla)
}
\arguments{
\item{fmla}{a formula}
}
\value{
the LHS as a character
}
\description{
Extract the LHS of a formula as a character
}
