% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{lssvd}
\alias{lssvd}
\title{svd version of lsfit}
\usage{
lssvd(x, y, zero = 10^(-16), has_intercept = all(cbind(x)[, 1] == 1))
}
\arguments{
\item{x}{predictor matrix, should include intercept term if needed}

\item{y}{response vector or matrix}

\item{zero}{value below which a latent value is considered to be 0}
}
\description{
Works with rank-deficient models (currently not centering)
}
