% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{qstats_resplot}
\alias{qstats_resplot}
\title{boxplot statistics for resplot}
\usage{
qstats_resplot(x, coef = 0.05, do.conf = TRUE, do.out = TRUE, ends = 0.05)
}
\description{
boxplot statistics for resplot
}
