% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{resplot}
\alias{resplot}
\title{Plot individual residuals}
\usage{
resplot(
  obj,
  data = obj$dout,
  log = FALSE,
  which = 1,
  at = seq(-2e+05, 1e+05, 10000),
  ...,
  rot = 45,
  par.strip.text = list(cex = 1)
)
}
\description{
Plot individual residuals
}
