% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{subrow}
\alias{subrow}
\title{Subtract selected row(s) of a matrix from all rows}
\usage{
subrow(L, i)
}
\arguments{
\item{L}{a matrix, typically a hypothesis matrix}

\item{i}{a vector used as row indices for L}
}
\value{
L minus the rows of L with index i, where the selected rows
         are repeated nrow(L)/length(i) times. For example, if
         L has 4 rows and i= 1, then the first row of L is subtracted from
         all rows. If i=c(1,3), then the first row is subtracted from
         rows 1 and 2, and the third row is subtracted from rows 3 and 4.
}
\description{
Subtract selected row(s) of a matrix from all rows
}
\examples{
L <- cbind(1, 1:4, c(-1,1,1,-1))
subrow(L, c(1,3))
}
