% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{gapplot}
\alias{gapplot}
\title{Plotting a decomp object}
\usage{
gapplot(
  obj,
  data = obj$gaps_each,
  log = FALSE,
  rot = 45,
  at = seq(-2e+05, 1e+05, 10000),
  auto.key = list(space = "right"),
  ...
)
}
\arguments{
\item{decomp}{a decomp object creaate by \code{\link{decomp}}.}
}
\description{
Plotting a decomp object
}
