% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icp.R
\name{icp}
\alias{icp}
\alias{aic}
\alias{bic}
\alias{ics}
\title{Graphical comparison of AIC and BIC}
\usage{
icp(ll, ..., rot = 0, srt = rot, prefix = "")

aic(ll, k = 2)

bic(ll, k = sapply(ll, function(f) log(nobs(f))))

ics(ll)
}
\arguments{
\item{ll}{a list of models using the same data.}

\item{rot, srt}{angle to rotate labels}

\item{prefix}{for labels}
}
\description{
Compares AIC and BIC for a named list of models. The models should have been fitted
to the same data.
}
\details{
Updated with version in UTFA
}
\section{Functions}{
\itemize{
\item \code{aic()}: compute AIC for a list of models.

\item \code{bic()}: compute BIC for a list of models.

\item \code{ics()}: compute BIC for a list of models

}}
\examples{
fit1 <- lm(mpg ~ wt, mtcars)
fit2 <- lm(mpg ~ wt + cyl, mtcars)
fit3 <- lm(mpg ~ wt * hp + cyl , mtcars)
fitl <- list('wt' = fit1, 'wt+cyl' = fit2, fit3=fit3)
aic(fitl)
ics(fitl)
icp(fitl)
}
