% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{lssvd}
\alias{lssvd}
\alias{lssvd_old}
\title{svd version of lsfit}
\usage{
lssvd(x, y, zero = 1e-07, ...)

lssvd_old(x, y, zero = 10^(-7), has_intercept = all(cbind(x)[, 1] == 1))
}
\arguments{
\item{x}{predictor matrix, should include intercept term if needed}

\item{y}{response vector or matrix}

\item{zero}{value below which a latent value of the data matrix is considered to be 0, default 10^(-7) in parallel with \code{\link{lsfit}}.}
}
\description{
Works with rank-deficient models (currently not centering)
}
\section{Functions}{
\itemize{
\item \code{lssvd_old}: oldversion
}}

